MMVv2: Flatten out abstract BasePcpWriter into PcpMmvWriter
BasePcpWriter only has a single implementation and I think it will make it
harder to implement MMVv2 support.
